+++ src/ProSuite.QA.Container/TestContainer/IUniqueIdProvider.cs
@@ -1,15 +1,15 @@
 using System.Collections.Generic;
-using ProSuite.Commons.AO.Geodatabase;
 
 namespace ProSuite.QA.Container.TestContainer
 {
 	public interface IUniqueIdProvider<T> : IUniqueIdProvider
 	{
-		int GetUniqueId(T feature);
+		long GetUniqueId(T feature);
 	}
+
 	public interface IUniqueIdProvider
 	{
-		bool Remove(int id);
+		bool Remove(long id);
 
 		IList<InvolvedRow> GetInvolvedRows(long id);
